Extract CacheFeedImageUseCaseTests to separate file
