+++ EssentialDevelopper/EssentialDeveloperFrameworkTests/Feed Cache/LoadFeedImageFromCacheUseCaseTests.swift
@@ -79,37 +79,6 @@ class LoadFeedImageFromCacheUseCaseTests: XCTestCase {
         store.complete(with: anyData(), at: 0)
     }
     
-    func test_save_triggersInsertInFeedImageStore() {
-        let (sut, store) = makeSUT()
-        let data = anyData()
-        let url = anyURL()
-        
-        sut.save(image: data, for: url) { _ in }
-        
-        XCTAssertEqual(store.messages, [.insert(url: url, data: data)])
-    }
-    
-    func test_save_deliversErrorOnInsertError() {
-        let (sut, store) = makeSUT()
-        let data = anyData()
-        let url = anyURL()
-        let exp = expectation(description: "wait for insert to complete")
-        let expectedError = anyNSError()
-        
-        sut.save(image: data, for: url) { result in
-            switch result {
-            case .failure:
-                break
-            case .success:
-                XCTFail("Expected to get error on insertion error got \(result) instead")
-            }
-            exp.fulfill()
-        }
-        store.insertComplete(with: expectedError)
-        wait(for: [exp], timeout: 1.0)
-    }
-    
-    
     private func expect(sut: LocalFeedImageLoader, toLoad expectedResult: FeedImageLoader.Result, when action: @escaping VoidClosure, file: StaticString = #file, line: UInt = #line) {
         
         let retreivedResult = result(from: sut, when: action)
